Add http trigger and helper class
